Update and unify services
