+++ src/ffs/service.py
@@ -11,12 +11,14 @@ import threading
 import time
 from datetime import datetime
 from pathlib import Path
-from typing import Iterator, Optional
-
+from typing import Iterator, Optional, Tuple
+import numpy as np
 import workflows.recipe
 from pydantic import BaseModel, ValidationError
 from rich.logging import RichHandler
 from workflows.services.common_service import CommonService
+import ffs.index
+from ssx_index import GPUIndexer
 
 logger = logging.getLogger(__name__)
 logger.level = logging.DEBUG
@@ -38,13 +40,14 @@ class PiaRequest(BaseModel):
     detector_distance: float
     d_min: float | None = None
     d_max: float | None = None
+    cell: tuple[float, float, float, float, float, float] | None = None
 
 
 class DetectorGeometry(BaseModel):
     pixel_size_x: float = 0.075  # Default value for Eiger
     pixel_size_y: float = 0.075  # Default value for Eiger
-    image_size_x: int = 4148 # Default value for Eiger
-    image_size_y: int = 4362 # Default value for Eiger
+    image_size_x: int = 2068 # Default value for Eiger
+    image_size_y: int = 2162 # Default value for Eiger
     distance: float
     beam_center_x: float
     beam_center_y: float
@@ -177,6 +180,13 @@ class GPUPerImageAnalysis(CommonService):
         )
         self._spotfinder_executable = _find_spotfinder()
         self._order_resolver = MessageOrderResolver(self.log)
+        ## Initialise the fast-feedback-indexer
+        self.indexer = None
+        try:
+            self.indexer = GPUIndexer()
+        except ModuleNotFoundError:
+            self.log.debug("ffbidx not found, has the fast-feedback-indexer module been built and sourced?")
+
 
     def gpu_per_image_analysis(
         self,
@@ -205,6 +215,18 @@ class GPUPerImageAnalysis(CommonService):
             self.log.warning(
                 f"Rejecting PIA request for {parameters.dcgid}/{parameters.message_index}({parameters.dcid}): Invalid detector parameters \n{e}"
             )
+
+        if self.indexer and parameters.cell and parameters.wavelength:
+            ## We have all we need to index, so make up to date models.
+            cell = gemmi.UnitCell(*parameters.cell)
+            self.indexer.cell = np.reshape(np.array(cell.orth.mat, dtype="float32"), (3,3)) ## Cell as an orthogonalisation matrix
+            self.indexer.panel = ffs.index.make_panel(
+                detector_geometry.distance,
+                detector_geometry.beam_center_x, detector_geometry.beam_center_y,
+                detector_geometry.pixel_size_x, detector_geometry.pixel_size_y,
+                detector_geometry.image_size_x, detector_geometry.image_size_y
+            )
+            self.indexer.wavelength = parameters.wavelength
 
         start_time = time.monotonic()
         self.log.info(
@@ -311,6 +333,12 @@ class GPUPerImageAnalysis(CommonService):
                 data["file-seen-at"] = time.time()
                 # XRC has one-based-indexing
                 data["file-number"] += 1
+                ## Do indexing
+                if (self.indexer and self.cell and self.panel and self.wavelength):
+                    xyzobs_px = np.array(data["spot_centers"])
+                    indexing_result = self.indexer.index(xyzobs_px)
+                    result = indexing_result.model_dump()
+                    ## FIXME where to send the result on to...
                 self.log.info(f"Sending: {data}")
                 rw.set_default_channel("result")
                 rw.send_to("result", data)
